# Optimize BitmapIter::next