Also, some minor cleanup to the benchmarking.

On my m1 mac, this seems pretty much faster across the board:

```
# Note: I only ran the iteration benchmarks
$ critcmp before after -t 2
group                                         after                                   before
-----                                         -----                                   ------
iteration/into_iter/census-income             1.00     22.8±1.52ms 289.4 MElem/sec    1.08     24.7±0.09ms 266.7 MElem/sec
iteration/into_iter/census-income_srt         1.00     19.6±0.10ms 296.5 MElem/sec    1.09     21.4±0.11ms 271.0 MElem/sec
iteration/into_iter/census1881_srt            1.00  1972.2±14.02µs 329.2 MElem/sec    1.08      2.1±0.02ms 303.8 MElem/sec
iteration/into_iter/weather_sept_85           1.00     43.4±0.33ms 282.9 MElem/sec    1.09     47.4±0.49ms 259.1 MElem/sec
iteration/into_iter/weather_sept_85_srt       1.00     52.4±0.41ms 293.1 MElem/sec    1.10     57.4±0.56ms 267.5 MElem/sec
iteration/into_iter/wikileaks-noquotes_srt    1.00   886.9±26.84µs 309.7 MElem/sec    1.05   935.0±26.49µs 293.8 MElem/sec
iteration/iter rev/census1881                 1.00      2.4±0.02ms 396.6 MElem/sec    1.06      2.6±0.02ms 373.2 MElem/sec
iteration/iter rev/census1881_srt             1.00      2.1±0.04ms 314.9 MElem/sec    1.03      2.1±0.02ms 306.2 MElem/sec
iteration/iter rev/wikileaks-noquotes         1.00    668.7±6.64µs 392.7 MElem/sec    1.08    721.1±3.61µs 364.2 MElem/sec
iteration/iter/census-income                  1.00     20.4±0.10ms 322.8 MElem/sec    1.10     22.4±0.11ms 294.1 MElem/sec
iteration/iter/census-income_srt              1.00     17.7±0.19ms 329.1 MElem/sec    1.10     19.5±0.36ms 297.9 MElem/sec
iteration/iter/census1881_srt                 1.00  1905.9±15.00µs 340.6 MElem/sec    1.08      2.1±0.01ms 315.3 MElem/sec
iteration/iter/weather_sept_85                1.00     39.3±0.89ms 312.1 MElem/sec    1.09     43.0±2.34ms 285.6 MElem/sec
iteration/iter/weather_sept_85_srt            1.00     47.4±0.64ms 324.3 MElem/sec    1.10     52.0±0.45ms 295.2 MElem/sec
iteration/iter/wikileaks-noquotes_srt         1.00   812.5±26.46µs 338.1 MElem/sec    1.07   871.8±11.29µs 315.1 MElem/sec
```